Transactions page refactoring
